fix(suite-native): show flagship device image in follow-device review when model is unreadable
