+++ suite-native/device/src/index.ts
@@ -11,6 +11,7 @@ export * from './components/ConnectDeviceAnimation';
 export * from './components/ConnectorImage';
 export * from './components/DeviceImage';
 export * from './components/ContinueOnTrezorScreenContent';
+export * from './components/FollowDeviceScreenContent';
 export * from './components/ConnectAndUnlockDeviceScreenContent';
 export * from './components/TurnOnAndUnlockDeviceScreenContent';
 export * from './components/DeviceDangerBanner';
